Refactor action-form and create-form to use base class
