+++ app/components/base/form/action-form/action-form.component.ts
@@ -1,31 +1,12 @@
-import { Component, Input, NgZone } from "@angular/core";
-import { FormGroup } from "@angular/forms";
-import { MdDialogRef } from "@angular/material";
-import { autobind } from "core-decorators";
-import { Observable } from "rxjs";
+import { Component, Input } from "@angular/core";
 
-import { BatchError } from "app/models";
+import { FormBase } from "../form-base";
 
 @Component({
     selector: "bex-action-form",
     templateUrl: "action-form.html",
 })
-export class ActionFormComponent {
-    /**
-     * Form group used in this form
-     */
-    @Input()
-    public formGroup: FormGroup;
-
-    /**
-     * Sidebar ref if the form is used in the sidebar.
-     * If provided this will add a add and close button option that will close the sidebar when the form is submitted.
-     */
-    @Input()
-    public dialogRef: MdDialogRef<any>;
-
-    @Input()
-    public submit: () => Observable<any>;
+export class ActionFormComponent extends FormBase {
 
     @Input()
     public submitText = "Proceed";
@@ -33,47 +14,10 @@ export class ActionFormComponent {
     @Input()
     public cancelText = "Cancel";
 
-    public loading = false;
-
-    public showErrorTroubleshoot = false;
-
-    public error: BatchError = null;
-
-    constructor(private zone: NgZone) {
-
-    }
     /**
      * Enabled if the formGroup is valid or there is no formGroup
      */
     public get submitEnabled() {
         return !this.formGroup || this.formGroup.valid;
-    }
-
-    @autobind()
-    public action(): Observable<any> {
-        this.loading = true;
-        const obs = this.submit();
-        obs.subscribe({
-            complete: () => {
-                this.loading = false;
-                this.error = null;
-                setTimeout(() => {
-                    this.cancel();
-                }, 1000);
-            },
-            error: (e: BatchError) => {
-                this.loading = false;
-                this.error = e;
-            },
-        });
-        return obs;
-    }
-
-    public cancel() {
-        if (this.dialogRef) {
-            this.zone.run(() => {
-                this.dialogRef.close();
-            });
-        }
     }
 }
