+++ app/components/base/form/create-form/create-form.component.ts
@@ -1,69 +1,17 @@
-import { Component, Input } from "@angular/core";
-import { FormGroup } from "@angular/forms";
-import { autobind } from "core-decorators";
-import { Observable } from "rxjs";
+import { Component } from "@angular/core";
 
-import { BatchError } from "app/models";
-import { SidebarRef } from "../../sidebar";
+import { FormBase } from "../form-base";
 
 @Component({
     selector: "bex-create-form",
     templateUrl: "create-form.html",
 })
-export class CreateFormComponent {
-    /**
-     * Form group used in this form
-     */
-    @Input()
-    public formGroup: FormGroup;
-
-    /**
-     * Sidebar ref if the form is used in the sidebar.
-     * If provided this will add a add and close button option that will close the sidebar when the form is submitted.
-     */
-    @Input()
-    public sidebarRef: SidebarRef<any>;
-
-    @Input()
-    public submit: () => Observable<any>;
-
-    public loading = false;
-
-    public showErrorTroubleshoot = false;
-
-    public error: BatchError = null;
-
-    @autobind()
-    public add(): Observable<any> {
-        this.loading = true;
-        const obs = this.submit();
-        obs.subscribe({
-            next: () => {
-                this.loading = false;
-                this.error = null;
-            },
-            error: (e: BatchError) => {
-                this.loading = false;
-                this.error = e;
-            },
-        });
-        return obs;
+export class CreateFormComponent extends FormBase {
+    public add() {
+        return this.performAction();
     }
 
-    @autobind()
-    public addAndClose(): Observable<any> {
-        const obs = this.add();
-        obs.subscribe(() => {
-            setTimeout(() => {
-                this.close();
-            }, 1000);
-        });
-        return obs;
-    }
-
-    public close() {
-        if (this.sidebarRef) {
-            this.sidebarRef.destroy();
-        }
+    public addAndClose() {
+        return this.performActionAndClose();
     }
 }
