+++ test/app/components/job/action/terminate/terminate-job-dialog.component.spec.ts
@@ -59,7 +59,7 @@ describe("TerminateJobDialogComponent ", () => {
     });
 
     it("Submit should call service and close the dialog", () => {
-        actionForm.action();
+        actionForm.performActionAndClose();
 
         expect(jobServiceSpy.terminate).toHaveBeenCalledTimes(1);
         expect(jobServiceSpy.terminate).toHaveBeenCalledWith("job-1", {});
@@ -68,7 +68,7 @@ describe("TerminateJobDialogComponent ", () => {
     it("Submit should call service and show error if fail", () => {
         component.jobId = "bad-job-id";
         fixture.detectChanges();
-        actionForm.action();
+        actionForm.performActionAndClose();
 
         expect(jobServiceSpy.terminate).toHaveBeenCalledTimes(1);
         expect(jobServiceSpy.terminate).toHaveBeenCalledWith("bad-job-id", {});
